refactor: move routing setup to its own filer
